+++ server/src/app.module.ts
@@ -1,12 +1,12 @@
 import { MiddlewareConsumer, Module, NestModule } from "@nestjs/common";
 import { AppController } from "./app.controller";
 import { AppService } from "./app.service";
-import { AuthModule } from "./auth/auth.module";
-import { RoomsModule } from "./rooms/rooms.module";
+import { RoomsModule } from "./modules/rooms/rooms.module";
+import { AuthModule } from "./modules/auth/auth.module";
 import { MongooseModule } from "@nestjs/mongoose";
 import { ConfigModule } from "@nestjs/config";
-import { PhaserModule } from "./phaser/phaser.module";
-import { NotificationModule } from "./notification/notification.module";
+import { PhaserModule } from "./modules/phaser/phaser.module";
+import { NotificationModule } from "./modules/notification/notification.module";
 import * as mongoose from "mongoose";
 import dotenv from "dotenv";
