--- server/src/auth/auth.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from "@nestjs/testing";
-import { AuthService } from "./auth.service";
-
-describe("AuthService", () => {
-    let service: AuthService;
-
-    beforeEach(async () => {
-        const module: TestingModule = await Test.createTestingModule({
-            providers: [AuthService],
-        }).compile();
-
-        service = module.get<AuthService>(AuthService);
-    });
-
-    it("should be defined", () => {
-        expect(service).toBeDefined();
-    });
-});
