+++ server/src/common/types/request.type.ts
@@ -1,0 +1,6 @@
+import { Request } from "express";
+import { User } from "src/modules/users/schemas/user.schema";
+
+export type LoginRequest = Request & {
+    user: User;
+};
