+++ server/src/modules/auth/strategies/jwt.strategy.ts
