+++ server/src/modules/phaser/phaser.module.ts
