+++ server/src/modules/rooms/dto/create-room.dto.ts
