+++ server/src/modules/rooms/dto/message-info.dto.ts
@@ -1,0 +1,5 @@
+export class MessageInfoDTO {
+    roomId: string;
+    message: string;
+    nickname: string;
+}
