+++ server/src/modules/rooms/dto/update-room.dto.ts
