--- server/src/rooms/dto/messageInfo.dto.ts
@@ -1,5 +1,0 @@
-export class MessageInfoDTO {
-    roomId : string;
-    message : string;
-    nickname : string;
-}
