--- server/src/types/request.type.ts
@@ -1,6 +1,0 @@
-import { Request } from "express";
-import { User } from "src/schema/user.schema";
-
-export type LoginRequest = Request & {
-    user: User;
-};
