--- server/src/users/users.module.ts
@@ -1,18 +1,0 @@
-import { Module, forwardRef } from "@nestjs/common";
-import { UsersController } from "./users.controller";
-import { UsersService } from "./users.service";
-import { MongooseModule } from "@nestjs/mongoose";
-import { User, UserSchema } from "src/schema/user.schema";
-import { UsersRepository } from "./users.repository";
-import { AuthModule } from "src/auth/auth.module";
-
-@Module({
-    imports: [
-        MongooseModule.forFeature([{ name: User.name, schema: UserSchema }]),
-        forwardRef(() => AuthModule),
-    ],
-    controllers: [UsersController],
-    providers: [UsersService, UsersRepository],
-    exports: [UsersRepository],
-})
-export class UsersModule {}
