--- server/src/users/users.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from "@nestjs/testing";
-import { UsersService } from "./users.service";
-
-describe("UsersService", () => {
-    let service: UsersService;
-
-    beforeEach(async () => {
-        const module: TestingModule = await Test.createTestingModule({
-            providers: [UsersService],
-        }).compile();
-
-        service = module.get<UsersService>(UsersService);
-    });
-
-    it("should be defined", () => {
-        expect(service).toBeDefined();
-    });
-});
